requirements.txt
python-dotenv
pypaperless==5.1.0
requests
pydantic
GitPython
pypdf